+++ changelog.md
@@ -1,10 +1,19 @@
 [TOC]
+# 1.1 branch #
+## 1.1.1-beta ##
+* Added control for downloading by cookies and database
+* Added database downloads tabel
+* Added update file for database update
+* Changed database sessions name
+
 # 1.0 branch #
 ## 1.0.6-beta ##
 * Mobile support added
 * Templates are optimized
+
 ## 1.0.5-beta ##
 * Suitable for binary prefix standards
+
 ## 1.0.4-beta ##
 * Added acp script
 * Added control for updates in acp
@@ -12,6 +21,7 @@
 * Corrected bug in template file browser
 * Fixed bugs in templates
 * Moved update script in acp
+
 ## 1.0.3-beta ##
 * Corrected 404 bug
 * Removed update bug